Replace icon cache algorithm with hash table
For a "normal" situation with 100-300 icons, the sorted array with
binary search was not performing well, due to O(n) insertion (memmoving)
and O(logn) lookups. This simple HT with open addressing works good
enough with observed average LF of about 50% and 1-2 comparisons per
lookup (compared to 5-8 for binsearch before, for typical cache size of
100-300 icons). The extra space wasted on null-ref pointers is
considered negligible.